The main changes  are done in PSO_structure _test
The position_update method was where the errors were coming from. If the elements of the vectors were individually added (lines 87 and 88) then this would result in a bad fit. Therefore, the position and velocities were changed to a vector that has vectors in it. i.e Vec[Vec[1],Vec[2],........]. This seemed to give a better fit and keep things general.